Slowed down hunger and thirst meters The hunger meter will now take 6 days to go from 100 to 0 The thirst meter will take 3 days Added variable for hardcore mode and function to toggle it. Need to add a button on the settings menu that toggles it and updates the depletion rate

+++ Cat-apps/src/com/hogg/catapps/cat/CatHunger.java
@@ -9,6 +9,7 @@
  * Contributors:
  * 		James Garner
  * 		Drew Burden
+ * 		Daniel Thomas
  * 
  * Copyright © 2012 Hogg Studios
  * All rights reserved.
@@ -21,7 +22,7 @@ public class CatHunger extends Meter {
 	// by moodPercentageVariable% of the incrementAmount
 	double moodPercentageVariable = 0.10; 
 	
-	public CatHunger(int initValue, int incAmount) {
+	public CatHunger(int initValue, double incAmount) {
 		super(initValue, incAmount);
 	}
 }
