+++ Cat-apps/src/com/hogg/catapps/cat/CatThirst.java
@@ -9,6 +9,7 @@
  * Contributors:
  * 		James Garner
  * 		Drew Burden
+ * 		Daniel Thomas
  * 
  * Copyright © 2012 Hogg Studios
  * All rights reserved.
@@ -21,7 +22,7 @@ public class CatThirst extends Meter {
 	// by moodPercentageVariable% of the incrementAmount
 	double moodPercentageVariable = 0.10; 	
 	
-	public CatThirst(int initValue, int incAmount) {
+	public CatThirst(int initValue, double incAmount) {
 		super(initValue, incAmount);
 	}
 }
